Update CLI tests
